Добавляет модули в проект
